requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
redis==5.0.3
python-decouple==3.8

decouple==0.0.7